+++ server/config.py
@@ -1,0 +1,2 @@
+MINVALUE = 1000000
+KEYS_TO_KEEP = []
